requirements.txt
fastapi
uvicorn
python-multipart
face_recognition
qdrant-client
numpy
python-dotenv
